Synced Android makefiles against LineageOS angler target
The angler target of LineageOS seems to build, so I used it to recreate
the 'hi6250' target provided by 'Meticulus'.

The idea is to get rid of lecacy defines from the previos CM and settle
builds with the LineageOS fork asap to allow less painful dev on the
Huawei P9 Lite.

Change-Id: I52c0b9924b20ea5ae1e28a0fd9beaf5a28d3fee1
Signed-off-by: Carsten Schlote <carsten.schlote@gmail.com>